backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
starlette==0.37.2

motor==3.5.1
pymongo==4.6.3
dnspython==2.6.1

python-dotenv==1.0.1
pydantic==2.7.4
email-validator==2.2.0
python-multipart==0.0.9

passlib==1.7.4
bcrypt==3.2.2

reportlab==4.2.0

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.7.2
joblib==1.4.2